packages/forge-connectors/pyproject.toml
[project]
name = "forge-connectors"
version = "0.1.0"
description = "Unified connector framework for Open Forge data platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Open Forge Team" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "pyarrow>=14.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "httpx>=0.27",
    "structlog>=24.0",
    "croniter>=2.0",
]

[project.optional-dependencies]
vault = ["hvac>=2.0"]
aws = ["boto3>=1.34"]
kubernetes = ["kubernetes>=28.0"]
warehouse = [
    "snowflake-connector-python>=3.6",
    "google-cloud-bigquery>=3.0",
    "redshift-connector>=2.0",
]
streaming = [
    "aiokafka>=0.10",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.2",
]
all = [
    "forge-connectors[vault,aws,kubernetes,warehouse,streaming,dev]",
]

[project.entry-points."forge.connectors"]
postgresql = "forge_connectors.connectors.sql.postgresql:PostgreSQLConnector"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge_connectors"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
